+++ src/roomsdb.ts
@@ -1,4 +1,5 @@
 import WebSocket from "ws";
+import players from "./playersdb";
 
 export const rooms = new Map<string, { users: WebSocket[] }>();
 export const playerInfo = new Map<WebSocket, { name: string; index: string }>();
@@ -27,42 +28,6 @@ export function createRoom(ws: WebSocket) {
   notifyRoomUpdate();
 }
 
-export function addUserToRoom(roomId: string, ws: WebSocket) {
-  const room = rooms.get(roomId);
-  if (room) {
-    room.users.push(ws);
-
-    if (room.users.length === 2) {
-      const gameId = generateRoomId();
-      room.users.forEach((user, idx) => {
-        user.send(
-          JSON.stringify({
-            type: "create_game",
-            data: JSON.stringify({
-              idGame: gameId,
-              idPlayer: `player_${idx + 1}`,
-            }),
-            id: 0,
-          }),
-        );
-      });
-      rooms.delete(roomId);
-    }
-
-    notifyRoomUpdate();
-  } else {
-    ws.send(
-      JSON.stringify({
-        type: "error",
-        data: JSON.stringify({
-          message: "Комната не найдена.",
-        }),
-        id: 0,
-      }),
-    );
-  }
-}
-
 export function getRoomsList() {
   return Array.from(rooms.entries()).map(([roomId, room]) => ({
     roomId,
@@ -84,11 +49,9 @@ export function notifyRoomUpdate() {
 
   console.log(roomData);
 
-  rooms.forEach((room) => {
-    room.users.forEach((user) => {
-      if (user.readyState === user.OPEN) {
-        user.send(roomData);
-      }
-    });
+  players.forEach(({ ws }) => {
+    if (ws.readyState === ws.OPEN) {
+      ws.send(roomData);
+    }
   });
 }
